+++ client/src/pages/form.vue
@@ -320,6 +320,7 @@ function generateJsonOutput(filedata = {}) {
           item.type == "expression" ||
           item.type == "file" ||
           item.type == "table" ||
+          item.type == "yaml" ||
           (item.type == "datetime") || 
           false;
         var outputValue = undefined;